+++ .github/CHANGELOG.md
@@ -234,6 +234,7 @@
 1. [PFD] Reimplement PFD with msfs-avionics framework for improved performance - @saschl (saschl#9432)
 1. [EIS] Implement display switching - @tracernz (Mike)
 1. [FMGC] Automatic airway waypoint intersection - @tracernz (Mike)
+1. [FMGC] Add runway-by-itself approaches - @tracernz (Mike)
 
 ## 0.7.0